cleaning file verify
